Restructurize project
Move oauth package into internal/github
